+++ tests/tests_api_v1_0/test_api_v1_0_UTC.py
@@ -44,7 +44,7 @@ class CRUDTestCase(APITestCase):
             resp.status_code, status.HTTP_201_CREATED, msg=resp.content
         )
         url_create_date = resp.data['datetime']
-        self.assertNotEqual(url_create_date, "2018-05-07T12:44:29.000000Z")
+        self.assertNotEqual(url_create_date, "2018-05-07T12:44:29Z")
 
     @override_settings(TIME_ZONE='UTC', USE_TZ=True)
     def test_utc_timezone_utc(self):
@@ -60,7 +60,7 @@ class CRUDTestCase(APITestCase):
             resp.status_code, status.HTTP_201_CREATED, msg=resp.content
         )
         url_create_date = resp.data['datetime']
-        self.assertEqual(url_create_date, "2018-05-07T12:44:29.000000Z")
+        self.assertEqual(url_create_date, "2018-05-07T12:44:29Z")
 
     @override_settings(TIME_ZONE='Europe/Paris', USE_TZ=False)
     def test_utc_tz_false(self):